Move pip marker related constant & function in their own module
It didn't seem related with the rest of the locations module
