+++ src/pip/_internal/utils/misc.py
@@ -32,11 +32,11 @@ from pip import __version__
 from pip._internal.exceptions import CommandError, InstallationError
 from pip._internal.locations import (
     running_under_virtualenv, site_packages, user_site, virtualenv_no_global,
-    write_delete_marker_file,
 )
 from pip._internal.utils.compat import (
     WINDOWS, console_to_str, expanduser, stdlib_pkgs, str_to_display,
 )
+from pip._internal.utils.marker_files import write_delete_marker_file
 from pip._internal.utils.typing import MYPY_CHECK_RUNNING
 
 if PY2:
